requirements.txt
markdown==3.4.1
Jinja2==3.1.2
PyYAML==6.0